coloriz.py: generalize assignColor() by removing hardcoded role index
Previously, assignColor() would move the color role to the fourth-
highest position in the role list. This works fine for servers where
the bot's top role is within those top four roles, but would make it
much harder for the bot to be implemented in other servers with
uncertain role setups.

+ getBotTopRoleNum()
~ assignColor()