botCommands.py: implement color set command functionality
colorSet() implements the functionality to run the color set command.
It properly parses the user's input and returns a message to be sent
by the bot, saying what color the user picked. If an error is raised,
relavent info is instead returned for the user.

parseRGB() makes use of regex to better parse a user's input of an
RGB triplet. This makes the UX slightly better by being more
forgiving for extra characters, such as commas or parentheses.

+ parseRGB()
+ colorSet()